+++ src/audio/AudioEngine.cpp
@@ -14,7 +14,9 @@
 
 namespace pixelroot32::audio {
 
-    AudioEngine::AudioEngine(const AudioConfig& config, const pixelroot32::platforms::PlatformCapabilities& caps) 
+    namespace platforms = pixelroot32::platforms;
+
+    AudioEngine::AudioEngine(const AudioConfig& config, const platforms::PlatformCapabilities& caps) 
         : config(config), capabilities(caps) {
         #ifdef ESP32
                 scheduler = std::unique_ptr<ESP32AudioScheduler>(new ESP32AudioScheduler(caps.audioCoreId, caps.audioPriority));